Move signup modal
Make it clear this is a modal that can be opened from anywhere, it's not
constrained to one part of the flow.
